pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "draconic-bot"
version = "2.2.0"
description = "Discord support bot for TrinityCore-based WoW private servers"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.4.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/VoxCore84/draconic-bot"
Issues = "https://github.com/VoxCore84/draconic-bot/issues"
